# feat: add StaticParsedLine