--- a/Cargo.toml
+++ b/Cargo.toml
@@ -135,10 +135,6 @@ name = "encoders"
 harness = false
 
 [[bench]]
-name = "line_parser"
-harness = false
-
-[[bench]]
 name = "mapper"
 harness = false