--- a/influxdb_line_protocol/Cargo.toml
+++ b/influxdb_line_protocol/Cargo.toml
@@ -6,9 +6,9 @@ edition = "2018"
 
 [dependencies] # In alphabetical order
 nom = "5.1.1"
-smallvec = "1.2.0"
 snafu = "0.6.2"
 observability_deps = { path = "../observability_deps" }
+ouroboros = "0.8.3"
 
 [dev-dependencies] # In alphabetical order
 test_helpers = { path = "../test_helpers" }