--- a/influxdb_line_protocol/Cargo.toml
+++ b/influxdb_line_protocol/Cargo.toml
@@ -11,4 +11,9 @@ observability_deps = { path = "../observability_deps" }
 ouroboros = "0.8.3"
 
 [dev-dependencies] # In alphabetical order
+criterion = "0.3.3"
 test_helpers = { path = "../test_helpers" }
+
+[[bench]]
+name = "line_parser"
+harness = false