--- a/influxdb_line_protocol/Cargo.toml
+++ b/influxdb_line_protocol/Cargo.toml
@@ -5,6 +5,7 @@ authors = ["Paul Dix <paul@pauldix.net>"]
 edition = "2018"
 
 [dependencies] # In alphabetical order
+arrayvec = "0.7"
 nom = "5.1.1"
 snafu = "0.6.2"
 observability_deps = { path = "../observability_deps" }