… Cargo.lock
… influxdb_line_protocol/src/vec.rs