In order to aggregate lines across requests it is necessary for them to own the data they refer to, so they can be stored in a collection, etc...

This rejigs the line splitting logic to return index ranges, instead of string slices, to then allow ouroboros to work its magic and construct StaticParsedLine that retain a reference to the originating `Arc<str>`. This formalation preserves the zero-copy semantics of the parser.

Unfortunately there is [an issue](https://github.com/servo/rust-smallvec/issues/217) with smallvec that causes it to not be correctly covariant. This PR therefore removes SmallVec, especially since moving the jemalloc I imagine the difference will be largely negligible and the parser is certainly not the ingest bottleneck at the moment. Smallvec v2, whenever it is released, should fix this bug.