@alamb 2021-05-11
@tustvold  is there a writeup somewhere that describes what you plan to do with this code (you mention something about "aggregating across requests")?

I find I can offer much better reviews of code if I understand the usecase (aka the *why* are you making this change) which I don't feel like I do quite yet.