@tustvold 2021-05-13
@alamb I'm struggling to reproduce your findings...

On main

```
running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 82 filtered out; finished in 0.00s

line_parser/all lines   time:   [880.93 us 881.55 us 882.31 us]                                  
                        thrpt:  [125.83 MiB/s 125.94 MiB/s 126.03 MiB/s]
```

On this branch

```
line_parser/all lines   time:   [869.83 us 870.20 us 870.63 us]                                  
                        thrpt:  [127.52 MiB/s 127.58 MiB/s 127.64 MiB/s]
                 change:
                        time:   [-1.5726% -1.4175% -1.3014%] (p = 0.00 < 0.05)
                        thrpt:  [+1.3185% +1.4378% +1.5977%]
                        Performance has improved.
```