@alamb 2021-05-13
I reran the benchmark (on a macbook pro, btw) and I still see a slowdown. 

Specifically, I ran:

On main @ ebea554e5a9fd591efdcb32880b67ce639150092 `cargo bench -- line_parser`

Then on this branch @ 3d210bb8459e2df4e3ca1d44261940eef212229d  `cargo bench -p influxdb_line_protocol -- line_parser`

```
line_parser/all lines   time:   [1.9917 ms 2.0127 ms 2.0337 ms]                                  
                        thrpt:  [54.591 MiB/s 55.162 MiB/s 55.742 MiB/s]
                 change:
                        time:   [+55.166% +57.180% +59.220%] (p = 0.00 < 0.05)
                        thrpt:  [-37.194% -36.378% -35.553%]
                        Performance has regressed.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high severe
```

🤔 

(amusingly my benchmark reports speeds almost 2x slower in absolute magnitude than your machine :) )