@tustvold 2021-05-13
@alamb I've pushed a branch https://github.com/influxdata/influxdb_iox/tree/owned-lines-jemalloc that uses jemalloc instead of the system allocator, could you give that a try?