@alamb 2021-05-13
> that uses jemalloc instead of the system allocator, could you give that a try? 
... trying ...