@e-dard 2021-05-13
Here is what I see:

**macOS intel i9**
```
⇒  critcmp main owned-lines
group                                                                          main                                     owned-lines
-----                                                                          ----                                     -----------
line_parser/all lines                                                          1.00  1219.6±13.87µs    91.0 MB/sec      1.54  1877.9±37.19µs    59.1 MB/sec
```

**macOS M1 (`aarch64`)
```
```

So for both of these there is a slow down. I think the linux behaviour is probably more important but we should get to the bottom of it because the regression on these architectures is so significant.