@alamb 2021-05-13
Here is the comparison of main to `owned-lines-jemalloc`:. It still shows a slowdown but much less pronounced (10% rather than 50%)

```
line_parser/all lines   time:   [1.3603 ms 1.3650 ms 1.3704 ms]                                  
                        thrpt:  [81.015 MiB/s 81.333 MiB/s 81.614 MiB/s]
                 change:
                        time:   [+11.125% +11.826% +12.716%] (p = 0.00 < 0.05)
                        thrpt:  [-11.281% -10.576% -10.011%]
                        Performance has regressed.
Found 7 outliers among 100 measurements (7.00%)
  5 (5.00%) high mild
  2 (2.00%) high severe
```