@tustvold 2021-05-16
I've pushed a commit that adds a replacement for Smallvec, or at least the functionality of it we need. Testing shows this to perform on-par with smallvec, but without the covariance issue. It does, however, use a small amount of unsafe - this makes about a ~8% difference to the benchmark, I will be interested to hear what people think about this.

On my Linux machine this branch now represents a pretty consistent speedup of ~2%. On mac the story is somewhat more complex. I'm testing on an Intel Mac Mini and the benchmark performance is rather inconsistent, it is common to see swings of 12% on the same code and I've struggled to find a consistent testing methodology - it almost seems like the system doesn't ramp up to fullspeed until it has seen a consistent load for a period, but then if it has seen a consistent load for too long it starts to throttle. The benchmark results are below, I'm inclined to call it probably equivalent.

Tests were run with `cargo bench` on this branch and `line-protocol-bench` (which moves the benchmark to avoid having to compile the server crate)

```
Mac

line-protocol-bench

line_parser/all lines   time:   [1.4093 ms 1.4163 ms 1.4244 ms]                                  
                        thrpt:  [77.945 MiB/s 78.388 MiB/s 78.777 MiB/s]


line_parser/all lines   time:   [1.4064 ms 1.4098 ms 1.4139 ms]                                  
                        thrpt:  [78.522 MiB/s 78.748 MiB/s 78.942 MiB/s]
                        
line_parser/all lines   time:   [1.2621 ms 1.2653 ms 1.2693 ms]                                  
                        thrpt:  [87.466 MiB/s 87.743 MiB/s 87.965 MiB/s]

line_parser/all lines   time:   [1.4104 ms 1.4171 ms 1.4245 ms]                                  
                        thrpt:  [77.936 MiB/s 78.344 MiB/s 78.718 MiB/s]

line_parser/all lines   time:   [1.4093 ms 1.4155 ms 1.4227 ms]                                  
                        thrpt:  [78.039 MiB/s 78.432 MiB/s 78.776 MiB/s]


owned-lines


line_parser/all lines   time:   [1.3666 ms 1.3706 ms 1.3750 ms]                                  
                        thrpt:  [80.742 MiB/s 81.004 MiB/s 81.237 MiB/s]

line_parser/all lines   time:   [1.3698 ms 1.3749 ms 1.3807 ms]                                  
                        thrpt:  [80.409 MiB/s 80.748 MiB/s 81.053 MiB/s]


line_parser/all lines   time:   [1.3727 ms 1.3787 ms 1.3852 ms]                                  
                        thrpt:  [80.151 MiB/s 80.529 MiB/s 80.878 MiB/s]


line_parser/all lines   time:   [1.3693 ms 1.3759 ms 1.3833 ms]                                  
                        thrpt:  [80.260 MiB/s 80.691 MiB/s 81.078 MiB/s]


Linux

line-protocol-bench

line_parser/all lines   time:   [872.83 us 873.34 us 874.13 us]                                  
                        thrpt:  [127.01 MiB/s 127.12 MiB/s 127.20 MiB/s]

line_parser/all lines   time:   [898.19 us 898.72 us 899.17 us]                                  
                        thrpt:  [123.47 MiB/s 123.53 MiB/s 123.61 MiB/s]

line_parser/all lines   time:   [897.22 us 897.84 us 898.39 us]                                  
                        thrpt:  [123.58 MiB/s 123.65 MiB/s 123.74 MiB/s]

owned-lines

line_parser/all lines   time:   [851.61 us 852.13 us 852.99 us]                                  
                        thrpt:  [130.16 MiB/s 130.29 MiB/s 130.37 MiB/s]

line_parser/all lines   time:   [870.45 us 870.58 us 870.69 us]                                  
                        thrpt:  [127.51 MiB/s 127.53 MiB/s 127.55 MiB/s]

line_parser/all lines   time:   [860.31 us 860.51 us 860.74 us]                                  
                        thrpt:  [128.98 MiB/s 129.02 MiB/s 129.05 MiB/s]
```