@tustvold 2021-05-17
Following on from today's meeting, kinesis/kafka producer support is not on the near term roadmap - so not going to look to get these changes in at this time